+++ tests/o3/cartesian_spherical_harmonics_test.py
@@ -3,6 +3,7 @@ import math
 import pytest
 import torch
 from e3nn import o3
+from e3nn.util.test import assert_auto_jitable
 
 
 def test_weird_call():
@@ -114,3 +115,16 @@ def test_recurrence_relation(float_tolerance, l):
     )
 
     assert (a - b).abs().max() < 100*float_tolerance
+
+
+def test_module():
+    l = o3.Irreps("0e + 1o + 3o")
+    normalize = True
+    normalization = 'integral'
+    sp = o3.SphericalHarmonics(l, normalize, normalization)
+    sp_jit = assert_auto_jitable(sp)
+    xyz = torch.randn(11, 3)
+    assert torch.allclose(
+        sp_jit(xyz),
+        o3.spherical_harmonics(l, xyz, normalize, normalization)
+    )