+++ ChangeLog.md
@@ -10,8 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `o3.TensorProduct`: is jit scriptable
 - `o3.TensorProduct`: also broadcast the `weight` argument
 - simple e3nn models can be saved/loaded with `torch.save()`/`torch.load()`
+- JITable `o3.SphericalHarmonics` module version of `o3.spherical_harmonics`
+- `in_place` option for `e3nn.util.jit` compilation functions
 
 ### Changed
+- in `soft_one_hot_linspace` the argument `base` is renamed into `basis`
+- `Irreps.slices()`, do `zip(irreps.slices(), irreps)` to retrieve the old behavior
 - `math.soft_one_hot_linspace` very small change in the normalization of `fourier` basis
 - `normalize2mom` is now a `torch.nn.Module`
 - rename arguments `set_ir_...` into `filter_ir_...`